+++ libsolidity/interface/CompilerStack.cpp
@@ -78,12 +78,13 @@
 #include <liblangutil/SemVerHandler.h>
 #include <liblangutil/SourceReferenceFormatter.h>
 
-
 #include <libsolutil/SwarmHash.h>
 #include <libsolutil/IpfsHash.h>
 #include <libsolutil/JSON.h>
 #include <libsolutil/Algorithms.h>
 #include <libsolutil/FunctionSelector.h>
+
+#include <libevmasm/Ethdebug.h>
 
 #include <boost/algorithm/string/replace.hpp>
 
@@ -1196,9 +1197,7 @@ Json CompilerStack::ethdebug() const
 {
 	solAssert(m_stackState >= AnalysisSuccessful, "Analysis was not successful.");
 	solAssert(!m_contracts.empty());
-	Json result = Json::object();
-	result["sources"] = sourceNames();
-	return result;
+	return evmasm::ethdebug::resources(sourceNames(), VersionString);
 }
 
 Json CompilerStack::ethdebug(std::string const& _contractName) const
@@ -1216,13 +1215,10 @@ Json CompilerStack::ethdebug(Contract const& _contract, bool _runtime) const
 	solAssert(m_stackState >= AnalysisSuccessful, "Analysis was not successful.");
 	solAssert(_contract.contract);
 	solUnimplementedAssert(!isExperimentalSolidity());
-	if (_runtime)
-	{
-		Json result = Json::object();
-		return result;
-	}
-	Json result = Json::object();
-	return result;
+	evmasm::LinkerObject const& object = _runtime ? _contract.runtimeObject : _contract.object;
+	std::shared_ptr<evmasm::Assembly> const& assembly = _runtime ? _contract.evmRuntimeAssembly : _contract.evmAssembly;
+	solAssert(sourceIndices().contains(_contract.contract->sourceUnitName()));
+	return evmasm::ethdebug::program(_contract.contract->name(), sourceIndices()[_contract.contract->sourceUnitName()], assembly.get(), &object);
 }
 
 bytes CompilerStack::cborMetadata(std::string const& _contractName, bool _forIR) const