+++ libyul/YulStack.cpp
@@ -30,6 +30,7 @@
 #include <libyul/optimiser/Suite.h>
 #include <libyul/YulControlFlowGraphExporter.h>
 #include <libevmasm/Assembly.h>
+#include <libevmasm/Ethdebug.h>
 #include <liblangutil/Scanner.h>
 #include <liblangutil/SourceReferenceFormatter.h>
 
@@ -273,7 +274,8 @@ YulStack::assembleWithDeployed(std::optional<std::string_view> _deployName)
 				{{m_charStream->name(), 0}}
 			);
 		}
-		creationObject.ethdebug["not yet implemented @ MachineAssemblyObject::ethdebug"] = true;
+		if (debugInfoSelection().ethdebug)
+			creationObject.ethdebug = evmasm::ethdebug::program(creationObject.assembly->name(), 0, creationObject.assembly.get(), creationObject.bytecode.get());
 
 		if (deployedAssembly)
 		{
@@ -381,17 +383,6 @@ Json YulStack::astJson() const
 	yulAssert(m_parserResult, "");
 	yulAssert(m_parserResult->hasCode(), "");
 	return  m_parserResult->toJson();
-}
-
-Json YulStack::ethdebug() const
-{
-	yulAssert(m_parserResult, "");
-	yulAssert(m_parserResult->hasCode(), "");
-	yulAssert(m_parserResult->analysisInfo, "");
-
-	Json result = Json::object();
-	result["sources"] = Json::array({m_charStream->name()});
-	return result;
 }
 
 Json YulStack::cfgJson() const