+++ test/cmdlineTests/ethdebug_eof_container_osaka/input.sol
@@ -4,4 +4,3 @@ pragma solidity >=0.0;
 contract C {
     function f() public {}
 }
-